provider: one Parquet writer and one blake3 helper, not four
Three copies of the same code, found while auditing why the cache layer is
larger than the work it does.

write_parquet_stream (version_cache) and write_parquet_atomic (rollup_cache)
were the same thirty lines -- same create_dir_all, same tmp+rename, same ZSTD
properties, same writer loop -- differing only in that one returned the blake3
digest and the other returned unit. file_blake3 existed verbatim in both
rollup_cache and export.rs. The tests then spelled out
Pin<Box<dyn Stream<Item = Result<RecordBatch, Error>> + Send>> longhand in four
places despite version_cache::BatchStream being exactly that type.

Keeps the digest-returning form as the single writer, in version_cache, which
is the module both caches already depend on. Callers with nothing to record
drop the digest; write_sidecar says why it does (a sidecar's identity is its
filename, which carries the digest of the SOURCE, not of the Parquet).

This is not merely tidying. The archived rollup design doc instructed "mirror
crates/provider/src/format_cache.rs exactly, in a new sibling module", and the
copies then drifted -- which is how the cache layer ended up with several
differently-shaped mechanisms for one job, each individually incomplete. Copies
that differ only in a return value are the ones that drift silently.

One behavioural note preserved as a comment: the temp file must NOT be named
{stem}.parquet.tmp, because a ListingTable filtering on .parquet would scan a
partial write. Both copies got this right; only one said so.

Copilot-Session: 5dd3bc2b-4f9e-4a4b-85e0-19618b9d1ad7
